% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_final_score.R
\name{get_all_score}
\alias{get_all_score}
\title{get all score}
\usage{
get_all_score(studyid, path_db, domain = c("lb", "mi", "bw"))
}
\arguments{
\item{studyid}{Mandatory, character \cr
Studyid number}

\item{path_db}{Mandatory, character \cr
path of database}

\item{domain}{Mandatory, character or vector \cr
either string or vector of domain name}
}
\value{
dataframe
}
\description{
get all score
}
\examples{
\dontrun{
get_all_score(studyid='1234123', path_db='path/to/database.db', domain=c('lb'))
}
}
