% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bw_score.R
\name{get_bw_score}
\alias{get_bw_score}
\title{get BW score for a given studyid}
\usage{
get_bw_score(
  studyid,
  path_db,
  fake_study = FALSE,
  master_CompileData = NULL,
  score_in_list_format = FALSE
)
}
\arguments{
\item{studyid}{Mandatory, character \cr
Studyid number}

\item{path_db}{Mandatory, character \cr
path of database}

\item{fake_study}{optional, Boolean \cr
whether study generated by SENDsanitizer package}
}
\value{
dataframe
}
\description{
get BW score for a given studyid
}
\examples{
\dontrun{
get_bw_score(studyid='1234123', path_db='path/to/database.db')
}
}
