% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compile_data.R
\name{get_compile_data}
\alias{get_compile_data}
\title{filter out tk and recovery animal}
\usage{
get_compile_data(studyid, path_db)
}
\arguments{
\item{studyid}{Mandatory, character \cr
Studyid number}

\item{path_db}{Mandatory, character \cr
path of database}
}
\value{
dataframe
}
\description{
filter out tk and recovery animal
}
\examples{
\dontrun{
get_compile_data(studyid='1234123', path_db='path/to/database.db')
}
}
