% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_doses.R
\name{get_doses}
\alias{get_doses}
\title{filter out tk and recovery animal}
\usage{
get_doses(studyid, path_db, xpt_dir = NULL)
}
\arguments{
\item{studyid}{Mandatory, character \cr
Studyid number}

\item{path_db}{Mandatory, character \cr
path of database}
}
\value{
dataframe
}
\description{
filter out tk and recovery animal
}
\examples{
\dontrun{
get_doses(studyid='1234123', path_db='path/to/database.db')
}
}
