% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lb_score.R
\name{get_lb_score}
\alias{get_lb_score}
\title{get LB score for a given studyid}
\usage{
get_lb_score(studyid, path_db, fake_study = FALSE)
}
\arguments{
\item{studyid}{Mandatory, character \cr
Studyid number}

\item{path_db}{Mandatory, character \cr
path of database}

\item{fake_study}{optional, Boolean \cr
whether study generated by SENDsanitizer package}
}
\value{
score
}
\description{
get LB score for a given studyid
}
\examples{
\dontrun{
get_lb_score(studyid='1234123', path_db='path/to/database.db')
}
}
