% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mi_score.R
\name{get_mi_score}
\alias{get_mi_score}
\title{get MI score for a given studyid}
\usage{
get_mi_score(studyid, path_db)
}
\arguments{
\item{studyid}{Mandatory, character \cr
Studyid number}

\item{path_db}{Mandatory, character \cr
path of database}
}
\value{
score
}
\description{
get MI score for a given studyid
}
\examples{
\dontrun{
get_mi_score(studyid='1234123', path_db='path/to/database.db')
}
}
