% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupSEND.R
\name{groupSEND}
\alias{groupSEND}
\title{Merge demographic and treatment data for each of the SEND experimental results domains}
\usage{
groupSEND(
  dataset,
  targetDomain,
  dmFields = c("SEX", "ARMCD", "SETCD", "USUBJID", "RFSTDTC"),
  exFields = c("EXTRT", "EXDOSE", "EXDOSU"),
  txParams = c("TRTDOS", "TRTDOSU", "TKDESC", "GRPLBL", "SPGRPCD")
)
}
\arguments{
\item{dataset}{Mandatory, character\cr
dataset}

\item{dmFields}{Mandatory, vector\cr
vector of variables}

\item{exFields}{Mandatory, vector\cr
vector of variables}
}
\value{
table
}
\description{
Merge demographic and treatment data for each of the SEND experimental results domains
}
\examples{
\dontrun{
groupSEND()
}
}
