% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossStudyVisualizationApp_LoadExternal.R
\name{send_cross_study_app}
\alias{send_cross_study_app}
\title{Run sendCrossStudy app}
\usage{
send_cross_study_app(database_path)
}
\arguments{
\item{database_path}{Optional, character\cr
file path for database}
}
\value{
function run the app.
}
\description{
Run sendCrossStudy app
}
\examples{
\dontrun{
send_cross_study_app()
}
}
