% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_new.R
\name{sendSummarizer_app}
\alias{sendSummarizer_app}
\title{Run sendSummarizer shiny app}
\usage{
sendSummarizer_app(db_path)
}
\arguments{
\item{db_path}{Optional, character\cr
file path for database}
}
\value{
function run the app.
}
\description{
Run sendSummarizer shiny app
}
\examples{
\dontrun{
sendSummarizer_app()
}
}
